move visitors in one file per class
